+++ src/Adapter/Fetch/Fetch.mjs
@@ -2,5 +2,5 @@
 /** @typedef {import("../Method/METHOD.mjs").METHOD} METHOD */
 
 /**
- * @typedef {{url: string, query_params?: {[key: string]: *} | null, method?: METHOD, data?: *, abort_signal?: AbortSignal, no_ui?: boolean, no_error_ui?: boolean, no_authentication_ui?: boolean, assert_type?: ASSERT_TYPE}} Fetch
+ * @typedef {{url: string, query_params?: {[key: string]: *} | null, method?: METHOD, data?: *, abort_signal?: AbortSignal, no_ui?: boolean, no_error_ui?: boolean, no_authenticate?: boolean, assert_type?: ASSERT_TYPE}} Fetch
  */